pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "erlc-api"
version = "2.0.0"
description = "Lightweight sync and async Python wrapper for the ER:LC PRC API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = ["httpx>=0.27.0"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
erlc_api = ["py.typed"]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "mypy>=1.8",
  "ruff>=0.3"
]
webhooks = ["cryptography>=42.0"]
export = ["openpyxl>=3.1"]
time = ["python-dateutil>=2.9"]
rich = ["rich>=13.0"]
scheduling = ["apscheduler>=3.10"]
utils = [
  "openpyxl>=3.1",
  "python-dateutil>=2.9",
  "rich>=13.0",
  "apscheduler>=3.10",
]
all = [
  "cryptography>=42.0",
  "openpyxl>=3.1",
  "python-dateutil>=2.9",
  "rich>=13.0",
  "apscheduler>=3.10",
]
